[BE] refactor : 각 클래스들을 폴더별 구조화하여 정리
associate with #17
