+++ BE/side-dish-1/src/main/java/com/codesquad/sidedish1/domain/entity/Token.java
@@ -1,4 +1,4 @@
-package com.codesquad.sidedish1.presentation;
+package com.codesquad.sidedish1.domain.entity;
 
 import org.springframework.data.annotation.Id;
