+++ BE/side-dish-1/src/main/java/com/codesquad/sidedish1/domain/repository/TokenRepository.java
@@ -1,5 +1,6 @@
-package com.codesquad.sidedish1.presentation;
+package com.codesquad.sidedish1.domain.repository;
 
+import com.codesquad.sidedish1.domain.entity.Token;
 import org.springframework.data.repository.CrudRepository;
 import org.springframework.stereotype.Repository;
