+++ BE/side-dish-1/src/main/java/com/codesquad/sidedish1/presentation/LoginController.java
@@ -1,5 +1,10 @@
 package com.codesquad.sidedish1.presentation;
 
+import com.codesquad.sidedish1.domain.entity.AccessTokenResponseSuccess;
+import com.codesquad.sidedish1.domain.entity.Token;
+import com.codesquad.sidedish1.domain.entity.User;
+import com.codesquad.sidedish1.domain.repository.TokenRepository;
+import com.codesquad.sidedish1.domain.repository.UserRepository;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 import org.springframework.beans.factory.annotation.Autowired;
