+++ BE/side-dish-1/src/main/java/com/codesquad/sidedish1/presentation/LoginService.java
@@ -1,5 +1,6 @@
 package com.codesquad.sidedish1.presentation;
 
+import com.codesquad.sidedish1.domain.entity.AccessTokenResponseSuccess;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 import org.springframework.beans.factory.annotation.Value;
